+++ database/migrations/2020_03_09_125744_create_car_people_table.php
@@ -16,9 +16,9 @@ class CreateCarPeopleTable extends Migration
         Schema::create('car_people', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('obs');
-            $table->integer('car_id')->unsigned();
-            $table->integer('person_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('car_id')->unsigned();
+            $table->bigInteger('person_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->foreign('car_id')->references('id')->on('cars')->onDelete('cascade');
             $table->foreign('person_id')->references('id')->on('people')->onDelete('cascade');
